+++ CHANGELOG.rst
@@ -21,8 +21,7 @@ master
 Changed
 ~~~~~~~
 
-- (`#21 <https://github.com/openscm/openscm-runner/pull/21>`_) Flexible start date for FaIR; build scmdata to emissions converter on openscm-runner
-
+- (`#21 <https://github.com/openscm/openscm-runner/pull/21>`_) Added flexible start date for FaIR and FaIR's scmdata to emissions converter
 
 v0.4.1 - 2020-10-06
 -------------------